removing console.log and adding mocks

+++ contracts/integrations/yearn/IYVault.sol
@@ -27,11 +27,11 @@ interface IYVault is IERC20 {
         uint256 maxLoss
     ) external returns (uint256);
 
-    function pricePerShare() external view returns (uint256);
+    function pricePerShare() external view  returns (uint256);
 
     function name() external view returns (string memory);
 
     function symbol() external view returns (string memory);
 
-    function decimals() external view returns (uint256);
+    function decimals() external view returns (uint8);
 }
